[INTERNAL] rename include/seqan3/alphabet/quality/{,quality_}composition.hpp
