+++ include/seqan3/alphabet/quality.hpp
@@ -35,7 +35,7 @@
 #pragma once
 
 #include <seqan3/alphabet/quality/aliases.hpp>
-#include <seqan3/alphabet/quality/composition.hpp>
+#include <seqan3/alphabet/quality/quality_composition.hpp>
 #include <seqan3/alphabet/quality/concept.hpp>
 #include <seqan3/alphabet/quality/illumina18.hpp>
