+++ include/seqan3/alphabet/quality/aliases.hpp
@@ -45,7 +45,7 @@
 #include <utility>
 
 #include <seqan3/alphabet/concept.hpp>
-#include <seqan3/alphabet/quality/composition.hpp>
+#include <seqan3/alphabet/quality/quality_composition.hpp>
 #include <seqan3/alphabet/quality/concept.hpp>
 #include <seqan3/alphabet/quality/illumina18.hpp>
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
