+++ include/seqan3/alphabet/quality/quality_composition.hpp
