+++ test/alphabet/quality/quality_composition_test.cpp
@@ -37,7 +37,7 @@
 #include <seqan3/alphabet/concept.hpp>
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/quality/illumina18.hpp>
-#include <seqan3/alphabet/quality/composition.hpp>
+#include <seqan3/alphabet/quality/quality_composition.hpp>
 
 using namespace seqan3;
